SegmentoPulse/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# News & RSS
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3

# HTTP client (replaces aiohttp)
httpx==0.26.0

# Caching (optional - can work without Redis)
# redis==5.0.1
# hiredis==2.3.2

# Firebase
firebase-admin==6.4.0

# Data processing
python-dateutil==2.8.2

# CORS & Security
python-multipart==0.0.6
email-validator==2.1.0
